+++ Adapters/AppLovin/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## AppLovin iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationAppLovin', '13.5.0.0'
+pod 'CASMediationAppLovin', '13.5.1.0'
 ```
+
+### 13.5.1.0
+- Certified with AppLovin - 13.5.1
+- Workaround for a crash caused by ALCASBridge.
 
 ### 13.5.0.0
 - Certified with AppLovin - 13.5.0