+++ Adapters/GoogleAds/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Google Ads iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationGoogleAds', '12.12.0.0'
+pod 'CASMediationGoogleAds', '12.14.0.0'
 ```
+
+### 12.14.0.0
+- Certified with GoogleAds - 12.14.0
 
 ### 12.12.0.0
 - Certified with GoogleAds - 12.12.0