+++ Adapters/Verve/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Maticoo iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationVerve', '3.7.0.0'
+pod 'CASMediationVerve', '3.7.1.0'
 ```
+
+### 3.7.1.0
+- Certified with Verve - 3.7.1
 
 ### 3.7.0.0
 - Certified with Verve - 3.7.0