+++ Adapters/YangoAds/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Yango Ads iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationYangoAds', '7.16.2.0'
+pod 'CASMediationYangoAds', '7.17.1.0'
 ```
+
+### 7.17.1.0
+- Certified with YangoAds - 7.17.1
+- Added creative id information.
 
 ### 7.16.2.0
 - Certified with YangoAds - 7.16.2